Bug 1902875 - Part 7: Add tests
