+++ tests/tests/files/field-layout/field-type-include.h
@@ -1,0 +1,9 @@
+  Type1 other_file_field1;
+
+  Container1<Type1>
+    other_file_field2;
+
+  std::vector<
+    Type1
+  >
+    other_file_field3;
