+++ tests/tests/files/field-layout/field-type.cpp
@@ -1,55 +1,7 @@
 #include <vector>
 #include <stddef.h>
 #include <stdint.h>
-
-class JSObject {
-};
-
-namespace JS {
-
-class TestAllocPolicy {
-};
-
-
-template <typename T>
-class Rooted {
-};
-
-template <typename T, size_t MinInlineCapacity = 0,
-          typename AllocPolicy = TestAllocPolicy>
-class GCVector {
-};
-
-template <typename T, typename AllocPolicy = TestAllocPolicy>
-class StackGCVector : public GCVector<T, 8, AllocPolicy> {
-};
-
-template <typename T>
-class RootedVector : public Rooted<StackGCVector<T>> {
-};
-
-class alignas(8) Value {
- private:
-  uint64_t asBits_;
-};
-
-
-template <typename Key, typename Value>
-class GCHashMap {
-};
-
-}
-
-namespace js {
-
-class TestCheck {
-};
-
-template <typename Check, typename T>
-class ProtectedDataNoCheckArgs {
-};
-
-}
+#include "field-type.h"
 
 namespace field_layout {
 
@@ -74,6 +26,13 @@ struct Container2 {
   T a;
 };
 
+#define DEFINE_FIELDS \
+  Type1 macro_fields_1; \
+  Enum1 macro_fields_2; \
+  Container1<Type1> macro_fields_3;
+
+#define SOME_ANNOTATION __attribute__((annotate("some_annotation")))
+
 struct S {
   Type1 value_field;
   const Type1* pointer_field;
@@ -83,7 +42,14 @@ struct S {
   JS::Rooted<JS::Value> rooted_field;
   JS::RootedVector<JS::Value> rooted_vec_field;
   JS::GCHashMap<JS::Value, JSObject*> hash_map_field;
-  js::ProtectedDataNoCheckArgs<js::TestCheck,  JS::Value> protected_field;
+  js::TestLockData<JS::Value> protected_field;
+  DEFINE_FIELDS
+  js::TestLockData<
+    Type1
+  >
+  multiline_field
+  SOME_ANNOTATION;
+#include "field-type-include.h"
 };
 
 S f() {
